env file for dynamic server hash

+++ scripts/run.js
@@ -2,17 +2,23 @@ const webpack = require('webpack');
 const webpackClientDevConfig = require('./../webpack/webpack.config.client.dev');
 const webpackServerDevConfig = require('./../webpack/webpack.config.server.dev');
 const handleWebpackStats = require('./../webpack/handle-stats');
+const replaceEntry = require('./../webpack/replace-entry');
 const spawn = require('child_process').spawn;
 const chalk = require('chalk');
 const path = require('path');
 
 var server = null;
+var _clientEntry = '';
 const refreshServer = (clientEntry = null, serverEntry = null) => {
-  if (!serverEntry) return;
-  if (server) server.kill();
+  if (clientEntry) _clientEntry = clientEntry;
+  if (!serverEntry || !_clientEntry) return;
+
   const serverPath = path.resolve(webpackServerDevConfig.output.path, serverEntry);
-  console.log(chalk.cyan(`${server ? 'Restarting' : 'Starting'} sambell...`));
-  server = spawn('node', [serverPath], { stdio: 'inherit', env: process.env });
+  replaceEntry(serverPath, _clientEntry, () => {
+    if (server) server.kill();
+    console.log(chalk.cyan(`${server ? 'Restarting' : 'Starting'} sambell...`));
+    server = spawn('node', [serverPath], { stdio: 'inherit', env: process.env });
+  });
 };
 
 webpack([
